Moved includeCache in his own file and made it a field of detector
